+++ Source/Modules/internal_plugins/internal_plugins.cpp
@@ -2,3 +2,4 @@
 #include "internal_plugins.h"
 
 #include "DrumSamplerPlugin/DrumSamplerPlugin.cpp"
+#include "DistortionPlugin/DistortionPlugin.cpp"